unexpose createOutgoingHttpHeaders from public api
