+++ packages/astro/src/core/app/node.ts
@@ -1,7 +1,8 @@
 import fs from 'node:fs';
 import { App } from './index.js';
 import { deserializeManifest } from './common.js';
-import type { IncomingMessage, ServerResponse, OutgoingHttpHeaders } from 'node:http';
+import { createOutgoingHttpHeaders } from './createOutgoingHttpHeaders.js';
+import type { IncomingMessage, ServerResponse } from 'node:http';
 import type { RouteData } from '../../@types/astro.js';
 import type { RenderOptions } from './index.js';
 import type { SerializedSSRManifest, SSRManifest } from './types.js';
@@ -186,36 +187,3 @@ export async function loadApp(rootFolder: URL): Promise<NodeApp> {
 	const manifest = await loadManifest(rootFolder);
 	return new NodeApp(manifest);
 }
-
-/**
- * Takes in a nullable WebAPI Headers object and produces a NodeJS OutgoingHttpHeaders object suitable for usage
- * with ServerResponse.writeHead(..) or ServerResponse.setHeader(..)
- *
- * @param headers WebAPI Headers object
- * @returns {OutgoingHttpHeaders} NodeJS OutgoingHttpHeaders object with multiple set-cookie handled as an array of values
- */
-export const createOutgoingHttpHeaders = (
-	headers: Headers | undefined | null
-): OutgoingHttpHeaders | undefined => {
-	if (!headers) {
-		return undefined;
-	}
-
-	// at this point, a multi-value'd set-cookie header is invalid (it was concatenated as a single CSV, which is not valid for set-cookie)
-	const nodeHeaders: OutgoingHttpHeaders = Object.fromEntries(headers.entries());
-
-	if (Object.keys(nodeHeaders).length === 0) {
-		return undefined;
-	}
-
-	// if there is > 1 set-cookie header, we have to fix it to be an array of values
-	if (headers.has('set-cookie')) {
-		const cookieHeaders = headers.getSetCookie();
-		if (cookieHeaders.length > 1) {
-			// the Headers.entries() API already normalized all header names to lower case so we can safely index this as 'set-cookie'
-			nodeHeaders['set-cookie'] = cookieHeaders;
-		}
-	}
-
-	return nodeHeaders;
-};
